implemented Symbolic Link

+++ nsh.sh
@@ -2013,7 +2013,7 @@ nsh() {
             draw_help_line "Press $NSH_COLOR_SH1 v \e[0m to edit the file using the default editor. To check the default editor, see NSH_DEFAULT_EDITOR in config.\e[K"
             draw_help_line "See the table of important keyboard shortcuts below:\e[K"
             local c='Less' && [[ $show_all -eq 0 ]] && c='All'
-            draw_shortcut_ml "F2" "Rename" "F5" "Copy" "F6" "Move" "F7" "Mkdir" "F10" "Config" "g" "Git" "y" "Yank" "r" "Refresh" "i" "Rename" "I" "Touch" "dd" "Delete" "m" "Mark" "'" "Bookmarks" ';' "Commands" "Tab" "View" ":" "Shell" "/" "Search" "^G" "Grep" "." "Show$c" "s" "Sort" "~" "Home" "2" "2048"
+            draw_shortcut_ml "F2" "Rename" "F5" "Copy" "F6" "Move" "F7" "Mkdir" "F10" "Config" "g" "Git" "y" "Yank" "r" "Refresh" "i" "Rename" "I" "Touch" "dd" "Delete" "m" "Mark" "'" "Bookmarks" 'L' "Ln" ';' "Commands" "Tab" "View" ":" "Shell" "/" "Search" "^G" "Grep" "." "Show$c" "s" "Sort" "~" "Home" "2" "2048"
             show_cursor
             get_key
             disable_line_wrapping
@@ -2426,16 +2426,17 @@ nsh() {
         local yfocus
         local yy
         local ylast_item
-        local op='cp'
-        [[ $1 == 'mv' ]] && op='mv'
+        local op="$1"
         yopen() {
             path="$(command cd "$1/"; pwd)"
             [[ $path == / ]] && path=
             move_cursor "1;$((list_width+1))"
-            if [[ $op == 'mv' ]]; then
-                printf '\e[30;42m\e[K| Move to: %s\e[0m' "${path/#$HOME\//$tilde/}"
+            if [[ $op == cp ]]; then
+                printf '\e[30;42m\e[K| Copy to: %s\e[0m' "$(print_filename "$path")"
+            elif [[ $op == mv ]]; then
+                printf '\e[30;42m\e[K| Move to: %s\e[0m' "$(print_filename "$path")"
             else
-                printf '\e[30;42m\e[K| Copy to: %s\e[0m' "${path/#$HOME\//$tilde/}"
+                printf '\e[30;42m\e[K| Symbolic link\e[0m'
             fi
             dirs=()
             files=()
@@ -2486,14 +2487,16 @@ nsh() {
         }
         draw_shortcut_yank() {
             move_cursor $LINES
-            if [[ $op == 'mv' ]]; then
-                draw_shortcut "P" "Paste " "D" "Delete" "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
+            if [[ $op == cp ]]; then
+                draw_shortcut "p" "Paste " "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
+            elif [[ $op == mv ]]; then
+                draw_shortcut "p" "Paste " "D" "Delete" "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
             else
-                draw_shortcut "P" "Paste " "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
+                draw_shortcut "L" "Make Link" "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
             fi
         }
 
-        if [ ${#selected[@]} -eq 0 ]; then
+        if [[ ${#selected[@]} -eq 0 && $op != ln ]]; then
             if [[ ${list[$focus]} != ".." ]]; then
                 selected[$focus]="$PWD/${list[$focus]}"
             else
@@ -2638,31 +2641,42 @@ nsh() {
                     fi
                     draw_shortcut_yank
                     ;;
-                'p')
-                    if [[ $op == 'mv' ]]; then
-                        nshmv "${selected[@]/#$PWD\//}" "$path"
-                    else
-                        nshcp "${selected[@]/#$PWD\//}" "$path"
-                    fi
-                    #command cd "$path"
-                    last_item="$path/${list[$focus]}"
-                    update
-                    return
-                    ;;
-                'd')
-                    if [[ $op == 'mv' ]]; then
-                        if [[ ${#selected[@]} -eq 1 ]]; then
-                            dialog "Delete ${selected[@]/#$HOME/$tilde}?" " Yes " " No "
+                p)
+                    if [[ $op == cp || $op == mv ]]; then
+                        if [[ $op == 'mv' ]]; then
+                            nshmv "${selected[@]/#$PWD\//}" "$path"
                         else
-                            dialog "Delete ${#selected[@]} files?" " Yes " " No "
+                            nshcp "${selected[@]/#$PWD\//}" "$path"
                         fi
-                        [[ $? -eq 0 ]] && for f in "${selected[@]}"; do
-                            rm -rf "$f"
-                        done
+                        #command cd "$path"
+                        last_item="$path/${list[$focus]}"
+                        update
+                        return
                     fi
-                    update
-                    return
                     ;;
+                d)
+                    if [[ $op == cp || $op == mv ]]; then
+                        if [[ $op == 'mv' ]]; then
+                            if [[ ${#selected[@]} -eq 1 ]]; then
+                                dialog "Delete ${selected[@]/#$HOME/$tilde}?" " Yes " " No "
+                            else
+                                dialog "Delete ${#selected[@]} files?" " Yes " " No "
+                            fi
+                            [[ $? -eq 0 ]] && for f in "${selected[@]}"; do
+                                rm -rf "$f"
+                            done
+                        fi
+                        update
+                        return
+                    fi
+                    ;;
+                L)
+                    if [[ $op == ln ]]; then
+                        last_item="$PWD/${ylist[$yfocus]%/}"
+                        ln -s "$path/${ylist[$yfocus]}" "${last_item##*/}"
+                        update
+                        return
+                    fi
             esac
         done
         update
@@ -3852,7 +3866,7 @@ nsh() {
                 echo "${bookmarks[$i]%% *}   ${val/#$HOME\//$tilde\/}"
             done; for ((i=$((${#visited[@]}-1)); i>=0; i--)); do
                 echo "    ${visited[$i]}"
-            done) | menu -h $max_lines --popup --header 'Key  Address' --footer $f --return-idx --searchable --key v 'echo edit')"
+            done) | menu -h $max_lines --popup --header 'Key  Address' --footer "$(draw_shortcut v Edit)" --return-idx --searchable --key v 'echo edit')"
             if [[ $opened == yes ]]; then
                 disable_line_wrapping >&2
                 hide_cursor >&2
@@ -4209,10 +4223,13 @@ nsh() {
                     fi
                     ;;
                 'y'|$'\e[15~') # cp
-                    yank
+                    yank cp
                     ;;
                 'd'|$'\e[17~')
-                    yank 'mv'
+                    yank mv
+                    ;;
+                L)
+                    yank ln
                     ;;
                 $'\e[12~'|$'\eOQ'|'i') # F2
                     f="${list[$focus]}"
